new processes for new files
