+++ packages/tools/package.json
@@ -53,7 +53,6 @@
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
     "esprima": "^4.0.1",
-    "fkill": "^7.1.0",
     "getopts": "^2.3.0",
     "glob": "^7.1.6",
     "handlebars": "^4.7.7",
@@ -68,7 +67,6 @@
     "postcss-combine-duplicated-selectors": "^10.0.2",
     "postcss-import": "^14.0.0",
     "properties-reader": "^2.2.0",
-    "ps-list": "^7.2.0",
     "puppeteer": "^8.0.0",
     "recursive-readdir": "^2.2.2",
     "rimraf": "^3.0.2",
